+++ sphinx/builders/linkcheck.py
@@ -80,6 +80,15 @@ class CheckExternalLinksBuilder(DummyBuilder):
             )
             warnings.warn(deprecation_msg, RemovedInSphinx80Warning, stacklevel=1)
 
+        if not self.config.linkcheck_distinguish_timeouts:
+            deprecation_msg = (
+                "The default value for 'linkcheck_distinguish_timeouts' will change "
+                "from `False` in Sphinx 7.3+ to `True`, meaning that timeouts for "
+                "HTTP requests will be reported using a distinct 'timeout' status. "
+                "See https://github.com/sphinx-doc/sphinx/issues/11868 for details."
+            )
+            warnings.warn(deprecation_msg, RemovedInSphinx80Warning, stacklevel=1)
+
     def finish(self) -> None:
         checker = HyperlinkAvailabilityChecker(self.config)
         logger.info('')
@@ -307,6 +316,7 @@ class HyperlinkAvailabilityCheckWorker(Thread):
         self.retries: int = config.linkcheck_retries
         self.rate_limit_timeout = config.linkcheck_rate_limit_timeout
         self._allow_unauthorized = config.linkcheck_allow_unauthorized
+        self._distinguish_timeouts = config.linkcheck_distinguish_timeouts
 
         self.user_agent = config.user_agent
         self.tls_verify = config.tls_verify
@@ -449,7 +459,7 @@ class HyperlinkAvailabilityCheckWorker(Thread):
                 break
 
             except RequestTimeout as err:
-                return 'timeout', str(err), 0
+                return 'timeout' if self._distinguish_timeouts else 'broken', str(err), 0
 
             except SSLError as err:
                 # SSL failure; report that the link is broken.
@@ -674,6 +684,7 @@ def setup(app: Sphinx) -> dict[str, Any]:
     app.add_config_value('linkcheck_anchors_ignore_for_url', (), '', (tuple, list))
     app.add_config_value('linkcheck_rate_limit_timeout', 300.0, '')
     app.add_config_value('linkcheck_allow_unauthorized', True, '')
+    app.add_config_value('linkcheck_distinguish_timeouts', False, '')
 
     app.add_event('linkcheck-process-uri')